+++ main.py
@@ -20,9 +20,9 @@ from TemperatureSensor import TemperatureSensor
 logging.basicConfig(level=logging.INFO)
 
 
-def get_bridge():
+def get_bridge(driver):
     """Call this method to get a Bridge instead of a standalone accessory."""
-    bridge = Bridge(display_name='Bridge')
+    bridge = Bridge(driver, display_name='Bridge')
     temp_sensor = TemperatureSensor('Sensor 2')
     temp_sensor2 = TemperatureSensor('Sensor 1')
     bridge.add_accessory(temp_sensor)
@@ -31,16 +31,18 @@ def get_bridge():
     return bridge
 
 
-def get_accessory():
+def get_accessory(driver):
     """Call this method to get a standalone Accessory."""
-    acc = TemperatureSensor('MyTempSensor')
+    acc = TemperatureSensor(driver, 'MyTempSensor')
     return acc
 
 
-acc = get_accessory()  # Change to get_bridge() if you want to run a Bridge.
-
 # Start the accessory on port 51826
-driver = AccessoryDriver(acc, port=51826)
+driver = AccessoryDriver(port=51826)
+
+acc = get_accessory(driver)  # Change to get_bridge(driver) if you want to run a Bridge.
+driver.add_accessory(acc)
+
 # We want KeyboardInterrupts and SIGTERM (kill) to be handled by the driver itself,
 # so that it can gracefully stop the accessory, server and advertising.
 signal.signal(signal.SIGINT, driver.signal_handler)